% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_hulls3d}
\alias{add_hulls3d}
\title{Plot hull 3d}
\usage{
add_hulls3d(x, polys, col = "red")
}
\arguments{
\item{x}{The output of the function plot used with a LAS object.}

\item{polys}{sf of sfc polygons returned by \link{find_bridges}}

\item{col}{color}
}
\description{
Function to display the bridges on the point cloud
}
