% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_bridge.R
\name{classify_bridges}
\alias{classify_bridges}
\title{Classify bridges using the output of \link{find_bridges}}
\usage{
classify_bridges(las, bbox, bottom_buffer = 0.75, upper_buffer = 2)
}
\arguments{
\item{las}{A LAS object.}

\item{bbox}{The bounding box found by \link{find_bridges}.}

\item{bottom_buffer}{Points below the bounding box are classified as bridge.}

\item{upper_buffer}{Points above the bounding box are classified as bridge. 2 m
in forest, 20 m in urban looks good.}
}
\description{
For an example see \link{find_bridges}.?
}
